i3plus: M300: enable Play Tone command
The LCD has a buzzer capable of producing a single tone, for a duration
between 10ms and 2.5s.  Enable usage of the M300 gcode command to play a
tone.  This is also be used elsewhere in code, such as in the M600 Filament
Change command which will be implemented soon.