i3plus: usb: add stop/pause/resume controls for USB prints
This adds support for stopping, pausing, and resuming USB prints from the
LCD panel.  The USB serial host will be notified of the event through the
standard "//action:" strings.  Note that the only standard way to notify the
host to stop a print is to send "//action:disconnect" so that it disconnects
from the serial port and cancels the print.

The UsbPrint page (which was the same as the Temperature page) is no longer
utilized, and the SdPrint page has been renamed to Print, as it is not used
for both print types.  Code was added to automatically switch between the
Print and Temperature pages depending on whether a print job was ongoing,
thereby showing the stop/pause/resume buttons only when they are applicable.

The temperature graphs are now also updated continuously from boot, instead
of from the first time the user opens the Temperature screen.  This allows
historical temperatures to be displayed, and prevents a 2 second delay after
opening the screen.

If the USB host does not process these actions commands, then the buttons
will not work.  OctoPrint has been verified to work with no additional
configuration.  The on-screen buttons may be used interchangibly with the
buttons on the USB host.

Using M125 to automatically pause like SD printing with ACTION_ON_PAUSE
defined causes OctoPrint to still run its "After print job is paused"
script, which interferes with normal operation of M125/M600.  Additionally,
resuming the print with M24 would always enable sdprinting.  Therefore, it
is simpler to let the host manage the pause, and this keeps the same
functionality for both on-screen buttons and buttons on the USB host.

Closes andrivet/ADVi3pp-Marlin#57