i3plus: load/unload: preserve preset hotend temperature
If the hotend temperature was already set before starting a load/unload,
then the screen should use that temperature as its default instead of 200.
Additionally, when exiting the screen, the hotend should not be disabled for
the cases where a print job is running (perhaps started while at this
screen) or the heater was enabled upon entering the menu.

This prevents accidently disabling the heater during a paused job and allows
for using screen menu to change filaments.